requirements.txt
#flask==1.1.2
#flask_restful
#pathlib
pandas
nltk
#numpy>=1.18.5
#pickle
PyYAML==5.3.1
scikit-learn
#scikit-learn==0.22.2.post1